backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
redis[asyncio]==5.0.4
pydantic-settings==2.2.1
httpx==0.27.0
websockets==12.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
